Restore unit doc and add tests on quickdoc
