+++ plugin/src/test/kotlin/ch/kleis/lcaac/plugin/language/ide/insight/LcaDocumentationProviderTest.kt
@@ -1,745 +1,252 @@
 package ch.kleis.lcaac.plugin.language.ide.insight
 
-import ch.kleis.lcaac.plugin.language.psi.LcaFile
-import ch.kleis.lcaac.plugin.psi.LcaDataRef
-import ch.kleis.lcaac.plugin.psi.LcaScaleQuantityExpression
-import com.intellij.psi.PsiManager
-import com.intellij.testFramework.fixtures.BasePlatformTestCase
-import com.jetbrains.rd.util.first
-import junit.framework.TestCase
+import ch.kleis.lcaac.plugin.language.ide.syntax.LcaCompletionTestCase
+import com.intellij.codeInsight.documentation.DocumentationManager
+import com.intellij.psi.util.PsiUtilBase
+import io.mockk.every
+import io.mockk.mockkObject
+import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.JUnit4
 
+private const val mockk_message = "ok_generator"
+
 @RunWith(JUnit4::class)
-class LcaDocumentationProviderTest : BasePlatformTestCase() {
-    override fun getTestDataPath(): String {
-        return ""
+class LcaDocumentationProviderTest : LcaCompletionTestCase() {
+    @Before
+    fun init() {
+        mockkObject(LcaDocumentGenerator)
     }
 
     @Test
-    fun testSubstance_ShouldRender_WithMinimumInfo() {
+    fun cursorOnUnitDeclaration_ShouldTriggerDoc() {
         // Given
-        val virtualFile = myFixture.createFile(
-            "testSubstance_ShouldRender_WithMinimumInfo.lca", """
-            substance co2 {
-                name = "co2"
-                type = Resource
-                compartment = "compartment"
-                reference_unit = kg
+        every { LcaDocumentGenerator.generateUnitDefinition(any()) } returns mockk_message
+        // When + Then
+        doTest(
+            getContent(Caret.UNIT_DECLARATION)
+        )
+    }
+
+    @Test
+    fun cursorOnUnitAlias_ShouldTriggerDoc() {
+        // Given
+        every { LcaDocumentGenerator.generateUnitDefinition(any()) } returns mockk_message
+        // When + Then
+        doTest(
+            getContent(Caret.UNIT_ALIAS)
+        )
+    }
+
+    @Test
+    fun cursorOnUnitInVariarableDeclaration_ShouldTriggerDoc() {
+        // Given
+        every { LcaDocumentGenerator.generateUnitDefinition(any()) } returns mockk_message
+        // When + Then
+        doTest(
+            getContent(Caret.UNIT_VAR_DECLARATION)
+        )
+    }
+
+    @Test
+    fun cursorOnUnitInQtyReference_ShouldTriggerDoc() {
+        // Given
+        every { LcaDocumentGenerator.generateUnitDefinition(any()) } returns mockk_message
+
+        // When + Then
+        doTest(
+            getContent(Caret.UNIT_QTY_REF)
+        )
+    }
+
+    @Test
+    fun cursorOnSubstanceDeclaration_ShouldTriggerDoc() {
+        // Given
+        every { LcaDocumentGenerator.generateSubstance(any()) } returns mockk_message
+
+        // When + Then
+        doTest(
+            getContent(Caret.SUBSTANCE_DECLARATION)
+        )
+    }
+
+    @Test
+    fun cursorOnSubstanceReference_ShouldTriggerDoc() {
+        // Given
+        every { LcaDocumentGenerator.generateSubstance(any()) } returns mockk_message
+
+        // When + Then
+        doTest(
+            getContent(Caret.SUBSTANCE_REF)
+        )
+    }
+
+
+    @Test
+    fun cursorOnProductDeclaration_ShouldTriggerDoc() {
+        // Given
+        every { LcaDocumentGenerator.generateProduct(any()) } returns mockk_message
+
+        // When + Then
+        doTest(
+            getContent(Caret.PROD_DECLARATION)
+        )
+    }
+
+    @Test
+    fun cursorOnProductReference_ShouldTriggerDoc() {
+        // Given
+        every { LcaDocumentGenerator.generateProduct(any()) } returns mockk_message
+
+        // When + Then
+        doTest(
+            getContent(Caret.PROD_REF)
+        )
+    }
+
+    @Test
+    fun cursorOnGlobalAssignmentDeclaration_ShouldTriggerDoc() {
+        // Given
+        every { LcaDocumentGenerator.generateGlobalAssignment(any()) } returns mockk_message
+
+        // When + Then
+        doTest(
+            getContent(Caret.GLOBAL_DECLARATION)
+        )
+    }
+
+    @Test
+    fun cursorOnGlobalAssignmentReference_ShouldTriggerDoc() {
+        // Given
+        every { LcaDocumentGenerator.generateGlobalAssignment(any()) } returns mockk_message
+
+        // When + Then
+        doTest(
+            getContent(Caret.GLOBAL_REF)
+        )
+    }
+
+    @Test
+    fun cursorOnLocalAssignmentDeclaration_ShouldTriggerDoc() {
+        // Given
+        every { LcaDocumentGenerator.generateAssignment(any()) } returns mockk_message
+
+        // When + Then
+        doTest(
+            getContent(Caret.LOCAL_DECLARATION)
+        )
+    }
+
+    @Test
+    fun cursorOnLocalAssignmentReference_ShouldTriggerDoc() {
+        // Given
+        every { LcaDocumentGenerator.generateAssignment(any()) } returns mockk_message
+
+        // When + Then
+        doTest(
+            getContent(Caret.LOCAL_REF)
+        )
+    }
+
+    @Test
+    fun cursorOnProcessDeclaration_ShouldTriggerDoc() {
+        // Given
+        every { LcaDocumentGenerator.generateProcess(any()) } returns mockk_message
+
+        // When + Then
+        doTest(
+            getContent(Caret.PROCESS_DECLARATION)
+        )
+    }
+
+    @Test
+    fun cursorOnProcessReference_ShouldTriggerDoc() {
+        // Given
+        every { LcaDocumentGenerator.generateProcess(any()) } returns mockk_message
+
+        // When + Then
+        doTest(
+            getContent(Caret.PROCESS_REF)
+        )
+    }
+
+
+    enum class Caret {
+        UNIT_DECLARATION, UNIT_ALIAS, UNIT_VAR_DECLARATION, UNIT_QTY_REF,
+        SUBSTANCE_REF, SUBSTANCE_DECLARATION,
+        PROD_REF, PROD_DECLARATION,
+        LOCAL_REF, LOCAL_DECLARATION,
+        GLOBAL_REF, GLOBAL_DECLARATION,
+        PROCESS_REF, PROCESS_DECLARATION
+    }
+
+    private fun getContent(car: Caret): String {
+        return """
+            package totoww
+
+            unit k${caret(car, Caret.UNIT_DECLARATION)}g2 {
+                symbol = "kgbis"
+                alias_for = 1.0001  k${caret(car, Caret.UNIT_ALIAS)}g
             }
-        """.trimIndent()
-        )
-        val file = PsiManager.getInstance(project).findFile(virtualFile) as LcaFile
-        val substance = file.getSubstances().first()
-        val sut = LcaDocumentationProvider()
 
-        // When
-        val result = sut.generateDoc(substance, substance)
-
-        // Then
-        TestCase.assertEquals(
-            """
-            <div class='definition'><pre>
-            <span style="color:#ffc800;font-style:italic;">Substance</span> <span style="color:#0000ff;font-weight:bold;">co2</span>
-            </pre></div>
-            <div class='content'>
-            <table class='sections'>
-            <tr>
-            <td valign='top' class='section'>Name</td>
-            <td valign='top'>co2</td>
-            </tr>
-            <tr>
-            <td valign='top' class='section'>Type</td>
-            <td valign='top'>Resource</td>
-            </tr>
-            <tr>
-            <td valign='top' class='section'>Compartment</td>
-            <td valign='top'>compartment</td>
-            </tr>
-            <tr>
-            <td valign='top' class='section'>Reference Unit</td>
-            <td valign='top'>kg</td>
-            </tr>
-            </table>
-            </div>
-            
-        """.trimIndent(), result
-        )
-    }
-
-    @Test
-    fun testSubstance_ShouldRender_WithAllInfos() {
-        // Given
-        val virtualFile = myFixture.createFile(
-            "testSubstance_ShouldRender_WithAllInfos.lca", """
-            substance propanol_air {
-                name = "propanol"
-                type = Resource
-                compartment = "air"
-                sub_compartment = "high altitude"
-                reference_unit = kg
-                meta{
-                    "author" = "Alain Colas"
-                    "description" = "Propan-1-ol...
-            with a return"
-                }
-            }
-        """.trimIndent()
-        )
-        val file = PsiManager.getInstance(project).findFile(virtualFile) as LcaFile
-        val substance = file.getSubstances().first()
-        val sut = LcaDocumentationProvider()
-
-        // When
-        val result = sut.generateDoc(substance, substance)
-
-        // Then
-        TestCase.assertEquals(
-            """
-        <div class='definition'><pre>
-        <span style="color:#ffc800;font-style:italic;">Substance</span> <span style="color:#0000ff;font-weight:bold;">propanol_air</span>
-        </pre></div>
-        <div class='content'>
-        <table class='sections'>
-        <tr>
-        <td valign='top' class='section'>Name</td>
-        <td valign='top'>propanol</td>
-        </tr>
-        <tr>
-        <td valign='top' class='section'>Type</td>
-        <td valign='top'>Resource</td>
-        </tr>
-        <tr>
-        <td valign='top' class='section'>Compartment</td>
-        <td valign='top'>air</td>
-        </tr>
-        <tr>
-        <td valign='top' class='section'>Sub-Compartment</td>
-        <td valign='top'>high altitude</td>
-        </tr>
-        <tr>
-        <td valign='top' class='section'>Reference Unit</td>
-        <td valign='top'>kg</td>
-        </tr>
-        </table>
-        </div>
-        <div class='content'>
-        <span style="">Propan-1-ol...<br>with a return</span></div>
-        <div class='content'>
-        <table class='sections'>
-        <tr>
-        <td valign='top' class='section'>Author</td>
-        <td valign='top'>Alain Colas</td>
-        </tr>
-        </table>
-        </div>
-        
-        """.trimIndent(), result
-        )
-    }
-
-    @Test
-    fun testQuantityRef_whenUnit_ShouldRender() {
-        // Given
-        val virtualFile = myFixture.createFile(
-            "testQuantityRef_whenUnit_ShouldRender.lca", """
-            process b {
-                params {
-                    yield = 100 kg
-                }
-            }
-        """.trimIndent()
-        )
-        val file = PsiManager.getInstance(project).findFile(virtualFile) as LcaFile
-        val assignment = file
-            .getProcesses().first()
-            .getParameters().first()
-            .value as LcaScaleQuantityExpression
-        val ref = assignment.dataExpression!! as LcaDataRef
-        val sut = LcaDocumentationProvider()
-
-        // When
-        val result = sut.generateDoc(ref, ref)
-
-        // Then
-        TestCase.assertEquals(
-            """
-        <div class='definition'><pre>
-        <span style="color:#ffc800;font-style:italic;">Unit</span> <span style="color:#0000ff;font-weight:bold;">kg</span>
-        </pre></div>
-        <div class='content'>
-        <table class='sections'>
-        <tr>
-        <td valign='top' class='section'>Symbol</td>
-        <td valign='top'>kg</td>
-        </tr>
-        <tr>
-        <td valign='top' class='section'>Dimension</td>
-        <td valign='top'>mass</td>
-        </tr>
-        </table>
-        </div>
-        
-        """.trimIndent(), result
-        )
-    }
-
-    @Test
-    fun testQuantityRef_whenUnitAlias_ShouldRender() {
-        // Given
-        val virtualFile = myFixture.createFile(
-            "testQuantityRef_whenUnitAlias_ShouldRender.lca", """
-            process b {
-                params {
-                    yield = 100 g
-                }
-            }
-        """.trimIndent()
-        )
-        val file = PsiManager.getInstance(project).findFile(virtualFile) as LcaFile
-        val assignment = file
-            .getProcesses().first()
-            .getParameters().first()
-            .value as LcaScaleQuantityExpression
-        val ref = assignment.dataExpression!! as LcaDataRef
-        val sut = LcaDocumentationProvider()
-
-        // When
-        val result = sut.generateDoc(ref, ref)
-
-        // Then
-        TestCase.assertEquals(
-            """
-        <div class='definition'><pre>
-        <span style="color:#ffc800;font-style:italic;">Unit</span> <span style="color:#0000ff;font-weight:bold;">g</span>
-        </pre></div>
-        <div class='content'>
-        <table class='sections'>
-        <tr>
-        <td valign='top' class='section'>Symbol</td>
-        <td valign='top'>g</td>
-        </tr>
-        <tr>
-        <td valign='top' class='section'>Alias for</td>
-        <td valign='top'>1e-3 kg</td>
-        </tr>
-        </table>
-        </div>
-        
-        """.trimIndent(), result
-        )
-    }
-
-    @Test
-    fun testQuantityRef_whenGlobalAssignment_ShouldRender() {
-        // Given
-        val virtualFile = myFixture.createFile(
-            "testQuantityRef_whenGlobalAssignment_ShouldRender.lca", """
             variables {
-                x = 1 kg
+                gl${caret(car, Caret.GLOBAL_DECLARATION)}o = 1 k${caret(car, Caret.UNIT_VAR_DECLARATION)}g
             }
-            
-            process b {
+
+            process fa${caret(car, Caret.PROCESS_DECLARATION)}rm {
+
                 params {
-                    yield = 100 x
+                    qt${caret(car, Caret.LOCAL_DECLARATION)}y = 1 g
                 }
-            }
-        """.trimIndent()
-        )
-        val file = PsiManager.getInstance(project).findFile(virtualFile) as LcaFile
-        val assignment: LcaScaleQuantityExpression = file
-            .getProcesses().first()
-            .getParameters().first()
-            .value as LcaScaleQuantityExpression
-        val ref = assignment.dataExpression!! as LcaDataRef
-        val sut = LcaDocumentationProvider()
 
-        // When
-        val result = sut.generateDoc(ref, ref)
-
-        // Then
-        TestCase.assertEquals(
-            """
-        <div class='definition'><pre>
-        <span style="color:#ffc800;font-style:italic;">Global quantity</span> <span style="color:#0000ff;font-weight:bold;">x</span>
-        </pre></div>
-        <div class='content'>
-        <table class='sections'>
-        <tr>
-        <td valign='top' class='section'>Symbol</td>
-        <td valign='top'>x</td>
-        </tr>
-        <tr>
-        <td valign='top' class='section'>Value</td>
-        <td valign='top'>1 kg</td>
-        </tr>
-        </table>
-        </div>
-        
-        """.trimIndent(), result
-        )
-    }
-
-    @Test
-    fun testQuantityRef_whenLocalVariable_ShouldRender() {
-        // Given
-        val virtualFile = myFixture.createFile(
-            "testQuantityRef_whenLocalVariable_ShouldRender.lca", """
-            process b {
-                variables {
-                    x = 1 kg
+                products {
+                   1 k${caret(car, Caret.UNIT_QTY_REF)}g car${caret(car, Caret.PROD_DECLARATION)}rot
                 }
-            
-                params {
-                    yield = 100 x
-                }
-            }
-        """.trimIndent()
-        )
-        val file = PsiManager.getInstance(project).findFile(virtualFile) as LcaFile
-        val assignment = file
-            .getProcesses().first()
-            .getParameters().first()
-            .value as LcaScaleQuantityExpression
-        val ref = assignment.dataExpression!! as LcaDataRef
-        val sut = LcaDocumentationProvider()
 
-        // When
-        val result = sut.generateDoc(ref, ref)
-
-        // Then
-        TestCase.assertEquals(
-            """
-        <div class='definition'><pre>
-        <span style="color:#ffc800;font-style:italic;">Quantity</span> <span style="color:#0000ff;font-weight:bold;">x</span>
-        </pre></div>
-        <div class='content'>
-        <table class='sections'>
-        <tr>
-        <td valign='top' class='section'>Symbol</td>
-        <td valign='top'>x</td>
-        </tr>
-        <tr>
-        <td valign='top' class='section'>Value</td>
-        <td valign='top'>1 kg</td>
-        </tr>
-        </table>
-        </div>
-        
-        """.trimIndent(), result
-        )
-    }
-
-    @Test
-    fun testQuantityRef_whenLocalParameter_ShouldRender() {
-        // Given
-        val virtualFile = myFixture.createFile(
-            "testQuantityRef_whenLocalParameter_ShouldRender.lca", """
-            process b {
-                params {
-                    x = 1 kg
-                }
-            
-                params {
-                    yield = 100 x
-                }
-            }
-        """.trimIndent()
-        )
-        val file = PsiManager.getInstance(project).findFile(virtualFile) as LcaFile
-        val assignment = file
-            .getProcesses().first()
-            .getParameters()["yield"]!! as LcaScaleQuantityExpression
-        val ref = assignment.dataExpression!! as LcaDataRef
-        val sut = LcaDocumentationProvider()
-
-        // When
-        val result = sut.generateDoc(ref, ref)
-
-        // Then
-        TestCase.assertEquals(
-            """
-        <div class='definition'><pre>
-        <span style="color:#ffc800;font-style:italic;">Quantity</span> <span style="color:#0000ff;font-weight:bold;">x</span>
-        </pre></div>
-        <div class='content'>
-        <table class='sections'>
-        <tr>
-        <td valign='top' class='section'>Symbol</td>
-        <td valign='top'>x</td>
-        </tr>
-        <tr>
-        <td valign='top' class='section'>Value</td>
-        <td valign='top'>1 kg</td>
-        </tr>
-        </table>
-        </div>
-        
-        """.trimIndent(), result
-        )
-    }
-
-    @Test
-    fun testDataRef_whenArgument_ShouldRender() {
-        // Given
-        val virtualFile = myFixture.createFile(
-            "testDataRef_whenArgument_ShouldRender.lca", """
-            process a {
                 inputs {
-                    1 kg carrot from b(x = 3 kg)
+                   3 g ca${caret(car, Caret.PROD_REF)}rrot from far${caret(car, Caret.PROCESS_REF)}m
+                   qt${caret(car, Caret.LOCAL_REF)}y salad
+                   gl${caret(car, Caret.GLOBAL_REF)}o salad
+                }
+
+                emissions {
+                    1.5 g myS${caret(car, Caret.SUBSTANCE_REF)}ubstance(compartment="air")
                 }
             }
-            
-            process b {
-                params {
-                    x = 1 kg
-                }
-                products {
-                    1 kg carrot
-                }
-            }
-        """.trimIndent()
-        )
-        val file = PsiManager.getInstance(project).findFile(virtualFile) as LcaFile
-        val ref = file.getProcesses().first()
-            .getInputs().first()
-            .inputProductSpec
-            .getProcessTemplateSpec()!!
-            .argumentList.first()
-            .parameterRef
-        val sut = LcaDocumentationProvider()
 
-        // When
-        val result = sut.generateDoc(ref, ref)
-
-        // Then
-        TestCase.assertEquals(
-            """
-        <div class='definition'><pre>
-        <span style="color:#ffc800;font-style:italic;">Quantity</span> <span style="color:#0000ff;font-weight:bold;">x</span>
-        </pre></div>
-        <div class='content'>
-        <table class='sections'>
-        <tr>
-        <td valign='top' class='section'>Symbol</td>
-        <td valign='top'>x</td>
-        </tr>
-        <tr>
-        <td valign='top' class='section'>Value</td>
-        <td valign='top'>1 kg</td>
-        </tr>
-        </table>
-        </div>
-        
-        """.trimIndent(), result
-        )
-    }
-
-    @Test
-    fun testProduct_ShouldRenderWithoutProcess() {
-        // Given
-        val virtualFile = myFixture.createFile(
-            "testProduct_ShouldRenderWithoutProcess.lca", """
-            process b {
-                labels {
-                    geo = "FR"
-                }
-                params {
-                    p1 = 1 kg
-                    p2 = p1 + p1
-                    p3 = "GLO"
-                }            
-                products {
-                    1 kg carrot
-                }
-            }
-        """.trimIndent()
-        )
-        val file = PsiManager.getInstance(project).findFile(virtualFile) as LcaFile
-        val carrotInputExchange = file
-            .getProcesses().first()
-            .getProducts().first()
-            .outputProductSpec
-        val sut = LcaDocumentationProvider()
-
-        // When
-        val result = sut.generateDoc(carrotInputExchange, carrotInputExchange)
-
-        // Then
-        TestCase.assertEquals(
-            """
-            <div class='definition'><pre>
-            <span style="color:#ffc800;font-style:italic;">Product</span> <span style="color:#0000ff;font-weight:bold;">carrot</span><span style="color:#ffc800;font-style:italic;"> from </span><span style="color:#0000ff;font-weight:bold;">b</span>
-            </pre></div>
-            <div class='content'>
-            <span style="color:#808080;font-style:italic;">Process Labels:</span><table class='sections'>
-            <tr>
-            <td valign='top' class='section'>geo = </td>
-            <td valign='top'>"FR"</td>
-            </tr>
-            </table>
-            </div>
-            <div class='content'>
-            <span style="color:#808080;font-style:italic;">Process Parameters:</span><table class='sections'>
-            <tr>
-            <td valign='top' class='section'>p1 = </td>
-            <td valign='top'>1 kg</td>
-            </tr>
-            <tr>
-            <td valign='top' class='section'>p2 = </td>
-            <td valign='top'>p1 + p1</td>
-            </tr>
-            <tr>
-            <td valign='top' class='section'>p3 = </td>
-            <td valign='top'>"GLO"</td>
-            </tr>
-            </table>
-            </div>
-            <div class='definition'><pre>
-            </pre></div>
-            
-        """.trimIndent(), result
-        )
-    }
-
-    @Test
-    fun testProcess_ShouldRenderWithoutProcess() {
-        // Given
-        val virtualFile = myFixture.createFile(
-            "testProcess_ShouldRenderWithoutProcess.lca", """
-            process b {
-                params {
-                    p1 = 1 kg
-                    p2 = p1 + p1
-                    p3 = "GLO"
-                }            
-                products {
-                    1 kg carrot
-                }
-            }
-        """.trimIndent()
-        )
-        val file = PsiManager.getInstance(project).findFile(virtualFile) as LcaFile
-        val process = file.getProcesses().first()
-        val sut = LcaDocumentationProvider()
-
-        // When
-        val result = sut.generateDoc(process, process)
-
-        // Then
-        TestCase.assertEquals(
-            """
-        <div class='definition'><pre>
-        <span style="color:#ffc800;font-style:italic;">Process</span> <span style="color:#0000ff;font-weight:bold;">b</span>
-        </pre></div>
-        <div class='content'>
-        <span style="color:#808080;font-style:italic;">Process Parameters:</span><table class='sections'>
-        <tr>
-        <td valign='top' class='section'>p1 = </td>
-        <td valign='top'>1 kg</td>
-        </tr>
-        <tr>
-        <td valign='top' class='section'>p2 = </td>
-        <td valign='top'>p1 + p1</td>
-        </tr>
-        <tr>
-        <td valign='top' class='section'>p3 = </td>
-        <td valign='top'>"GLO"</td>
-        </tr>
-        </table>
-        </div>
-        <div class='definition'><pre>
-        </pre></div>
-
-        """.trimIndent(), result
-        )
-    }
-
-    @Test
-    fun testProcess_ShouldRenderWithProcessAndMeta() {
-        // Given
-        val virtualFile = myFixture.createFile(
-            "testProcess_ShouldRenderWithProcessAndMeta.lca", """
-            process b {
+            substance mySub${caret(car, Caret.SUBSTANCE_DECLARATION)}stance {
+                name = "mySubstanceName"
+                type = Emission
+                compartment = "air"
+                reference_unit = kg2
                 meta {
-                    "author" = "Alain Colas"
-                    "description" = "Propan-1-ol..."
+                    "description" = "My doc"
+                    "author" = "My name"
                 }
-                params {
-                    p1 = 1 kg
-                    p2 = p1 + p1
-                }            
-                products {
-                    1 kg carrot
+
+                impacts {
+                    48.218 u Climate_change
                 }
             }
         """.trimIndent()
-        )
-        val file = PsiManager.getInstance(project).findFile(virtualFile) as LcaFile
-        val process = file.getProcesses().first()
-        val sut = LcaDocumentationProvider()
-
-        // When
-        val result = sut.generateDoc(process, process)
-
-        // Then
-        TestCase.assertEquals(
-            """
-        <div class='definition'><pre>
-        <span style="color:#ffc800;font-style:italic;">Process</span> <span style="color:#0000ff;font-weight:bold;">b</span>
-        </pre></div>
-        <div class='content'>
-        <span style="">Propan-1-ol...</span></div>
-        <div class='content'>
-        <table class='sections'>
-        <tr>
-        <td valign='top' class='section'>Author</td>
-        <td valign='top'>Alain Colas</td>
-        </tr>
-        </table>
-        </div>
-        <div class='content'>
-        <span style="color:#808080;font-style:italic;">Process Parameters:</span><table class='sections'>
-        <tr>
-        <td valign='top' class='section'>p1 = </td>
-        <td valign='top'>1 kg</td>
-        </tr>
-        <tr>
-        <td valign='top' class='section'>p2 = </td>
-        <td valign='top'>p1 + p1</td>
-        </tr>
-        </table>
-        </div>
-        <div class='definition'><pre>
-        </pre></div>
-
-        """.trimIndent(), result
-        )
     }
 
-    @Test
-    fun testProcess_ShouldRenderWithLabels() {
-        // Given
-        val virtualFile = myFixture.createFile(
-            "testProcess_ShouldRenderWithProcessAndMeta.lca", """
-            process b {
-                labels {
-                    geo = "GLO"
-                    env = "PROD"
-                }
-                meta {
-                    "author" = "Alain Colas"
-                    "description" = "Propan-1-ol..."
-                }
-                params {
-                    p1 = 1 kg
-                    p2 = p1 + p1
-                }            
-                products {
-                    1 kg carrot
-                }
-            }
-        """.trimIndent()
+    private fun caret(current: Caret, required: Caret) = if (current == required) "<caret>" else ""
+
+    private fun doTest(content: String) {
+        fixture.configureByText("sample.lca", content)
+        val originalElt = PsiUtilBase.getElementAtCaret(fixture.editor)
+        val element = DocumentationManager.getInstance(myFixture!!.project).findTargetElement(
+            myFixture!!.editor,
+            myFixture!!.file
         )
-        val file = PsiManager.getInstance(project).findFile(virtualFile) as LcaFile
-        val process = file.getProcesses().first()
-        val sut = LcaDocumentationProvider()
-
-        // When
-        val result = sut.generateDoc(process, process)
-
-        // Then
-        TestCase.assertEquals(
-            """
-        <div class='definition'><pre>
-        <span style="color:#ffc800;font-style:italic;">Process</span> <span style="color:#0000ff;font-weight:bold;">b</span>
-        </pre></div>
-        <div class='content'>
-        <span style="color:#808080;font-style:italic;">Process Labels:</span><table class='sections'>
-        <tr>
-        <td valign='top' class='section'>geo = </td>
-        <td valign='top'>"GLO"</td>
-        </tr>
-        <tr>
-        <td valign='top' class='section'>env = </td>
-        <td valign='top'>"PROD"</td>
-        </tr>
-        </table>
-        </div>
-        <div class='content'>
-        <span style="">Propan-1-ol...</span></div>
-        <div class='content'>
-        <table class='sections'>
-        <tr>
-        <td valign='top' class='section'>Author</td>
-        <td valign='top'>Alain Colas</td>
-        </tr>
-        </table>
-        </div>
-        <div class='content'>
-        <span style="color:#808080;font-style:italic;">Process Parameters:</span><table class='sections'>
-        <tr>
-        <td valign='top' class='section'>p1 = </td>
-        <td valign='top'>1 kg</td>
-        </tr>
-        <tr>
-        <td valign='top' class='section'>p2 = </td>
-        <td valign='top'>p1 + p1</td>
-        </tr>
-        </table>
-        </div>
-        <div class='definition'><pre>
-        </pre></div>
-
-        """.trimIndent(), result
-        )
+        val provider = DocumentationManager.getProviderFromElement(element)
+        kotlin.test.assertNotNull(provider)
+        val text = provider.generateDoc(element, originalElt)
+        kotlin.test.assertNotNull(text)
+        assertEquals(mockk_message, text)
     }
 
-    @Test
-    fun test_technoExchangeWithAllocateField_ShouldRender() {
-        // Given
-        val virtualFile = myFixture.createFile(
-            "test_technoExchangeWithAllocateField.lca", """
-            process glyphosate {
-                products {
-                    1 kg glyphosate allocate 90 percent
-                    1 kg waste allocate 10 percent
-                }
-            }
-            
-            process p {
-                products {
-                    1 kg p
-                }
-                inputs {
-                    1 kg glyphosate
-                }
-            }
-        """.trimIndent()
-        )
-        val file = PsiManager.getInstance(project).findFile(virtualFile) as LcaFile
-        val ref = file
-            .getProcesses().first()
-            .getProducts().first()
-            .outputProductSpec
-        val sut = LcaDocumentationProvider()
-        // When
-        val actual = sut.generateDoc(ref, ref)
-        // Then
-        TestCase.assertEquals(
-            """
-            <div class='definition'><pre>
-            <span style="color:#ffc800;font-style:italic;">Product</span> <span style="color:#0000ff;font-weight:bold;">glyphosate</span><span style="color:#ffc800;font-style:italic;"> from </span><span style="color:#0000ff;font-weight:bold;">glyphosate</span>
-            </pre></div>
-            <div class='content'>
-            <span style="color:#808080;font-style:italic;">Process Parameters:</span><table class='sections'>
-            </table>
-            </div>
-            <div class='definition'><pre>
-            </pre></div>
-            
-            """.trimIndent(), actual
-        )
-    }
+
 }
